+++ app/containers/OrderPage/OrderDetailPage/Loadable.ts
@@ -3,4 +3,4 @@
  */
 import loadable from 'loadable-components';
 
-export default loadable(() => import('./index'));
+export default loadable(() => import('.'));
