+++ app/containers/SideMenu/index.tsx
@@ -13,40 +13,39 @@ import { Link } from 'react-router-dom';
 import Sider from 'antd/lib/layout/Sider';
 import styled from 'styled-components';
 
-/* eslint-disable react/prefer-stateless-function */
-export default class SideMenu extends React.PureComponent {
-  public render() {
-    const Logo = styled.div`
+export default function SideMenu() {
+
+  const Logo = styled.div`
       height: 32px;
       background: ;
       margin: 16px;
       width: 250px;
     `;
 
-    return (
-      <Sider
-        breakpoint="lg"
-        theme="light"
-        collapsedWidth="0"
-        onBreakpoint={broken => {
-          console.log(broken);
-        }}
-        onCollapse={(collapsed, type) => {
-          console.log(collapsed, type);
-        }}
-      >
-        <Affix>
-          <Logo>
-            <img
-              src={require('../../images/covetrus_logo.png')}
-              alt="Test"
-              style={{ width: '150px' }}
-            />
-          </Logo>
-          {/* Set default key to store value */}
-          <Menu mode="inline" defaultSelectedKeys={['1']}>
-            <Menu.Item key="1">
-              <Link to="/">
+  return (
+    <Sider
+      breakpoint="lg"
+      theme="light"
+      collapsedWidth="0"
+      onBreakpoint={broken => {
+        console.log(broken);
+      }}
+      onCollapse={(collapsed, type) => {
+        console.log(collapsed, type);
+      }}
+    >
+      <Affix>
+        <Logo>
+          <img
+            src={require('../../images/covetrus_logo.png')}
+            alt="Test"
+            style={{ width: '150px' }}
+          />
+        </Logo>
+        {/* Set default key to store value */}
+        <Menu mode="inline" defaultSelectedKeys={['1']}>
+          <Menu.Item key="1">
+            <Link to="/">
                 <span>
                   <Icon
                     type="dashboard"
@@ -54,9 +53,9 @@ export default class SideMenu extends React.PureComponent {
                   />
                   Dashboard
                 </span>
-              </Link>
-            </Menu.Item>
-            {/* <Menu.Item key="2">
+            </Link>
+          </Menu.Item>
+          {/* <Menu.Item key="2">
               <Link to="/scripts">
                 <Icon
                   type="edit"
@@ -65,17 +64,17 @@ export default class SideMenu extends React.PureComponent {
                 Scripts
               </Link>
             </Menu.Item> */}
-            <Menu.Item key="3">
-              <Link to="/orders">
-                <Icon
-                  type="credit-card"
-                  theme="twoTone"
-                  style={{ display: 'inline-flex' }}
-                />
-                Orders
-              </Link>
-            </Menu.Item>
-            {/* <Menu.Item key="4">
+          <Menu.Item key="3">
+            <Link to="/orders">
+              <Icon
+                type="credit-card"
+                theme="twoTone"
+                style={{ display: 'inline-flex' }}
+              />
+              Orders
+            </Link>
+          </Menu.Item>
+          {/* <Menu.Item key="4">
               <Link to="/products">
                 <Icon
                   type="shop"
@@ -84,7 +83,7 @@ export default class SideMenu extends React.PureComponent {
                 Products
               </Link>
             </Menu.Item> */}
-            {/* <Menu.Item key="5">
+          {/* <Menu.Item key="5">
               <Link to="/exceptions">
                 <Icon
                   type="warning"
@@ -93,7 +92,7 @@ export default class SideMenu extends React.PureComponent {
                 Exceptions
               </Link>
             </Menu.Item> */}
-            {/* <Menu.Item key="6">
+          {/* <Menu.Item key="6">
               <Link to="/autoship">
                 <Icon
                   type="schedule"
@@ -102,9 +101,8 @@ export default class SideMenu extends React.PureComponent {
                 Autoship
               </Link>
             </Menu.Item> */}
-          </Menu>
-        </Affix>
-      </Sider>
-    );
-  }
+        </Menu>
+      </Affix>
+    </Sider>
+  );
 }
