+++ package.json
@@ -68,7 +68,7 @@
   },
   "dependencies": {
     "@babel/polyfill": "7.0.0",
-    "antd": "^3.15.0",
+    "antd": "^3.15.1",
     "chalk": "^2.4.2",
     "compression": "1.7.3",
     "connected-react-router": "4.5.0",
@@ -85,8 +85,8 @@
     "lodash": "4.17.11",
     "minimist": "1.2.0",
     "prop-types": "15.6.2",
-    "react": "^16.8.3",
-    "react-dom": "^16.8.3",
+    "react": "^16.8.4",
+    "react-dom": "^16.8.4",
     "react-helmet": "5.2.0",
     "react-intl": "2.7.2",
     "react-redux": "5.0.7",
@@ -117,12 +117,12 @@
     "@types/expect": "^1.20.4",
     "@types/express": "^4.16.1",
     "@types/fontfaceobserver": "0.0.6",
-    "@types/hoist-non-react-statics": "^3.0.1",
+    "@types/hoist-non-react-statics": "^3.3.0",
     "@types/invariant": "^2.2.28",
     "@types/jest": "^23.3.14",
-    "@types/lodash": "^4.14.121",
+    "@types/lodash": "^4.14.123",
     "@types/mocha": "^5.2.6",
-    "@types/node": "^10.12.27",
+    "@types/node": "^10.14.1",
     "@types/prop-types": "^15.7.0",
     "@types/react": "^16.8.6",
     "@types/react-dom": "^16.8.2",
@@ -134,10 +134,10 @@
     "@types/redux-devtools-dock-monitor": "^1.1.30",
     "@types/redux-devtools-log-monitor": "^1.0.30",
     "@types/sinon": "^5.0.7",
-    "@types/styled-components": "^4.1.11",
+    "@types/styled-components": "^4.1.12",
     "@types/systemjs": "^0.20.2",
     "@types/webpack": "^4.4.25",
-    "@types/webpack-env": "^1.13.7",
+    "@types/webpack-env": "^1.13.9",
     "add-asset-html-webpack-plugin": "3.1.1",
     "awesome-typescript-loader": "^5.2.0",
     "babel-core": "7.0.0-bridge.0",
@@ -187,7 +187,7 @@
     "rimraf": "2.6.2",
     "shelljs": "^0.8.3",
     "style-loader": "0.23.1",
-    "stylelint": "9.6.0",
+    "stylelint": "^9.10.1",
     "stylelint-config-recommended": "2.1.0",
     "stylelint-config-styled-components": "0.1.1",
     "stylelint-processor-styled-components": "1.5.0",
@@ -197,11 +197,11 @@
     "ts-node": "^7.0.0",
     "tsconfig-paths": "^3.8.0",
     "tsconfig-paths-webpack-plugin": "^3.2.0",
-    "tslint": "^5.13.0",
+    "tslint": "^5.14.0",
     "tslint-loader": "^3.4.3",
     "tslint-react": "^2.3.0",
     "typescript": "^3.3.3333",
-    "typescript-plugin-styled-components": "^1.2.0",
+    "typescript-plugin-styled-components": "^1.3.0",
     "url-loader": "1.1.2",
     "webpack": "4.23.1",
     "webpack-cli": "3.1.2",
